+++ CHANGELOG.rst
@@ -11,15 +11,23 @@ Patch releases 0.14.\*
 Next Release
 """"""""""""
 
+
 - Features:
-    - :issue_gh:`290` paste state at current mouse position (both via context menu and shortcut) @CSuerig
-    - add state at context menu position when using context menu to add states @CSuerig
 
 
 - Bug Fixes:
 
 
 - Miscellaneous:
+
+
+
+0.14.3
+""""""
+
+- Features:
+    - :issue_gh:`290` paste state at current mouse position (both via context menu and shortcut) @CSuerig
+    - add state at context menu position when using context menu to add states @CSuerig
 
 
 0.14.2